Move TarantoolConfigSpec to the right directory
